Set the type attribute of the buttons for logging out of all sessions to buttons
By not setting this attribute, the buttons default to submit and as the first
submit button on the page, it causes pressing the enter key to log you out of
all sessions. This change restores the pre 4.1 behavior where pressing enter
while focused on a form field submits the form.

merges [31010] to the 4.1 branch

props ocean90

fixes #30871


Built from https://develop.svn.wordpress.org/branches/4.1@31076