WP_Filesystem: Change `WP_Filesystem_FTPext::exists()` and `WP_Filesystem_ftpsockets::exists()` to return true for empty directories.
Both methods are using *nlist() which returns a list of files in a given directory or the file itself for a given file. If the result was an empty list we assumed that the file doesn't exists. This includes also cases where $file is actually an empty directory. To prevent this we now check if $file is a directory before returning the result of an empty list.
Other filesystem methods are using `file_exists()` which already checks whether a file or directory exists.

fixes #30815.
Built from https://develop.svn.wordpress.org/trunk@31815