About: Remove period from placeholder.
The position of periods is different in some languages, like Japanese. This removes the period from the last placeholder for the Template Hierarchy section. It doesn't include a new period to avoid a string change in a minor release.

Merge of [34054] to the 4.3 branch.

Props extendwings, chriscct7.
See #33429.
Built from https://develop.svn.wordpress.org/branches/4.3@34055