Build/Test Tools: Add path detection when running GitHub Actions on `pull_request`.
This change adds a list of paths to check when determining which workflows to run on `pull_request`. This will prevent certain workflows from running unnecessarily when the files updated are not related to the tests and checks being performed.

Props peterwilsoncc, johnbillion.
Fixes #52667.
Built from https://develop.svn.wordpress.org/trunk@50479