Block Editor: Load the classic layout stylesheet conditionallty.
Historically, the block editor has been opinionated about the size of the canvas and the alignments.
Themes with theme.json will have to define a "layout" in their settings and the classic stylesheet won't be loaded anymore for this.

Props nosolosw.
See #53175.

Built from https://develop.svn.wordpress.org/trunk@50992