Editor: Ensure the Query block pattern category is translatable.
Because the `__()` translation function was used here, the Query block pattern category string is not translatable. `_x()` is the correct function here as that passes the second parameter as context and not text domain.

Follow up to [50948].

Props walbo.
Merges [51307] to the 5.8 branch.
Fixes #53577.
Built from https://develop.svn.wordpress.org/branches/5.8@51308