Themes: Remove inline comment that is not relevant in WordPress Core.
This removes an inline comment block that was added when importing the `WP_Theme_JSON` class in [50959]. The comment is relevant to the code only when viewed within the context of the Gutenberg plugin.

Props nosolosw, jorbin, SergeyBiryukov.
Merges [51352] to the 5.8 branch.
Fixes #53576.
Built from https://develop.svn.wordpress.org/branches/5.8@51364