Site Editor: Fix typo in home template translatable description.
The home template's description should be: "Displays as the site's home page, or as the Posts page when a static home page isn't set." This commit fixes the typo changing "it" to "isn't".

Follow-up to [52331].

Props kharisblank, talldanwp.
Merges [52554] to the 5.9 branch.
Fixes #54787.
Built from https://develop.svn.wordpress.org/branches/5.9@52561