Media: Align username and post title in Save postbox.
Remove the float alignment that places the "Uploaded by" username and the "Uploaded to" post title on the opposite side of the Save postbox for mobile. This matches the alignment on desktop, and makes it easier for low-vision users to track the information relationship between label and content.

Props sabernhardt.
Fixes #55508.
Built from https://develop.svn.wordpress.org/trunk@53115